Fix wrapping for longer cell content
The previous fix worked for table cells that were not wrapped in
by "table-cell" markers, but didn't fix situations where the cell
content was more than 30 characters (and wrapped by table-cell
markers).

This patch moves the logic from handleText() to handleTableCell().
A utility function was added to calculate the total length of
content in the cell (taking into account that inline markup will
create "child nodes").

Checking for empty cells at the end of a row is no longer needed
because of this, as we can unconditionally add a cariage return
for the last cell (except for long lines, as table-cell markers
already have a newline included).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>